#15 adding unit tests for `towardsHome` and `towardsAway`
Adding unit tests for the Board View that test the functions
`towardsHome` and `towardsAway`. To do so (and because it would be
benificial) I have the tested functions returning a line number where
the ball should be at. Also the `moveChip` function now calculates
based on an initial Y value and the current line number