issue #35
I goofed! I did this on master instead of a branch

fixed the win state. Now you have to go to the options menu to start
over.

I also fixed the bug where the computer doesn't go after I score. There
still needs to be a timer on it though.